+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [**BREAKING**] a `default_code` expression can now refer to the values of earlier fields by name.
   (This is extremely unlikely to break your code, but could in theory due to shadowing.)
 - `#[builder(exclude)]` on fields, to not provide a method to set that field.
+- Control of documentation:
+  - `#[builder(doc = "…")]` on fields, to document the field’s method on the builder. Unlike `#[doc]`, you can currently only have one value rather than one attribute per line; but that’s not a big deal since you don’t get to use the `///` sugar anyway. Just use a multiline string.
+  - `#[builder(doc, builder_method_doc = "…", builder_type_doc = "…", build_method_doc = "…")]` on structs:
+    - `doc` unhides the builder type from the documentation.
+	- `builder_method_doc = "…"` replaces the default documentation that will be generated for the builder() method of the type for which the builder is being generated.
+	- `builder_type_doc = "…"` replaces the default documentation that will be generated for the builder type. Implies `doc`.
+	- `build_method_doc = "…"` replaces the default documentation that will be generated for the build() method of the builder type. Implies `doc`.
 
 ## 0.3.0 - 2019-02-19
 ### Added